Move config validation into config.py
